% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{invfn}
\alias{invfn}
\title{numerically invert a function}
\usage{
invfn(fn, ...)
}
\arguments{
\item{fn}{a function to be inverted}

\item{...}{additional parameters passed to \code{uniroot}}
}
\value{
an inverse function of \code{fn}
}
\description{
numerically invert a function
}
