% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{transfvoomxy}
\alias{transfvoomxy}
\title{transform mean-variance on log-cpm scale to count scale}
\usage{
transfvoomxy(voom.xy, logbase = 2)
}
\arguments{
\item{voom.xy}{the voom.xy component in the object returned from limma::voom}

\item{logbase}{base for the logorithm in \code{voom.xy}}
}
\value{
list of transformed mean and dispersion
}
\description{
transform the voom.xy component (log2-count, sqrt of residual standard deviation on log2-cpm)
in the object returned from limma::voom to mean and dispersion on the count scale
}
