% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{vstrans}
\alias{vstrans}
\alias{ivstrans}
\title{variance stabilizing transformation (VST) similar as DEXSeq}
\usage{
vstrans(a0, a1, meth = c("default", "DEXSeq"))

ivstrans(a0, a1, meth = c("default", "DEXSeq", "numeric"),
  vstfn = NULL, ...)
}
\arguments{
\item{a0, a1}{coefficients for the mean and dispersion relationship (d ~ a0 + a1/m),
as fitted via DESeq2:::parametricDispersionFit}

\item{meth}{method for deriving the analytical form of the transformation}

\item{vstfn}{a VST function if \code{meth} is 'numeric'}

\item{...}{additional parameters passed to \code{invfn}}
}
\value{
VST function with orginal count as input and return transformed count

inverse VST function with transformed count as input and return original
        untransformed count
}
\description{
variance stabilizing transformation (VST) similar as DEXSeq

inverse variance stabilizing transformation (VST)
}
